Move the "init" process to Init::main()
